rttanalysis: add rttanalysisccl.TestBenchmarkExpectations
Previously, we had no convenient way to adjust the roundtrip benchmark
expectations in the CCL package. This commit adds a new test for this
purpose and in so doing exports a bunch of test utilities in
rttanalysis.

Release note: None
