+++ pkg/ccl/benchccl/rttanalysisccl/BUILD.bazel
@@ -22,6 +22,7 @@ go_test(
     srcs = [
         "bench_test.go",
         "multi_region_bench_test.go",
+        "validate_benchmark_data_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":rttanalysisccl"],
@@ -33,6 +34,7 @@ go_test(
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/testutils/serverutils",
+        "//pkg/testutils/skip",
         "//pkg/testutils/testcluster",
         "//pkg/util/randutil",
     ],
